django-project/gymtracker/pyproject.toml
[tool.black]
line-length = 120
py36 = true
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''
[tool.poetry]
name = "gymtracker"
version = "0.1.0"
description = ""
authors = ["mairaafreitas <maira.oliveirafreitas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
djangorestframework = "*"
django-model-utils = "*"
pre-commit = "^3.0.4"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
